Refactor tutorial layout, add browser examples.
